+++ docs/changelog.rst
@@ -12,6 +12,8 @@ Breaking changes
   Explicit (:jira:`591`, :pull:`7`). By `Kyle Brindley`_.
 - Change Abaqus stress-type vector contraction function name to match the general purpose nature for both Standard and
   Explicit (:jira:`591`, :pull:`7`). By `Kyle Brindley`_.
+- Change Abaqus NTENS matrix contraction function name to match the general purpose nature for both Standard and
+  Explicit (:jira:`593`, :pull:`9`). By `Kyle Brindley`_.
 
 0.0.1 (2021-04-13)
 ------------------