+++ docs/changelog.rst
@@ -11,7 +11,9 @@ Breaking changes
 - Change Abaqus stress-type vector expansion function name to match the general purpose nature for both Standard and
   Explicit (:jira:`591`, :pull:`7`). By `Kyle Brindley`_.
 - Change Abaqus stress-type vector contraction function name to match the general purpose nature for both Standard and
-  Explicit (:jira:`591`, :pull:`7`). By `Kyle Brindley`_.
+  Explicit (:jira:`592`, :pull:`8`). By `Kyle Brindley`_.
+- Change Abaqus NTENS matrix contraction function name to match the general purpose nature for both Standard and
+  Explicit (:jira:`593`, :pull:`9`). By `Kyle Brindley`_.
 
 0.0.1 (2021-04-13)
 ------------------